+++ CHANGES.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 0.167
+- Upgrade RetroArch to Version 0.167
+- Added cores
+  - atari800-libretro
+  - prosystem-libretro
+  - stella-libretro
+  - fceumm_libretro  
+- Upgraded cores
+  - all
+- Removed cores
+  - pcsx_rearmed
+
+
 ## 0.163
 - Upgrade RetroArch to Version 0.163
 - Override Video and Input Drivers to prevent user selecting unsupported driver